🚚 rename: task 관련 schema, util함수 분리 (Jira Ticket : CCC-66)
