+++ components/modal-template/MakeTodoModal/index.tsx
@@ -14,7 +14,7 @@ import {
   FormLabel,
   FormMessage,
 } from '@/components/ui/form';
-import { todoModalFormSchema } from '@/lib/schema';
+import { todoModalFormSchema } from '@/lib/schema/task';
 import { zodResolver } from '@hookform/resolvers/zod';
 import React from 'react';
 import { Controller, useForm } from 'react-hook-form';
