+++ components/ui/toggle-group.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import { toggleVariants } from '@/components/ui/toggle';
-import cn from '@/lib/utils';
+import { cn } from '@/lib/utils';
 import * as ToggleGroupPrimitive from '@radix-ui/react-toggle-group';
 import { type VariantProps } from 'class-variance-authority';
 import * as React from 'react';
