+++ components/ui/toggle.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import cn from '@/lib/utils';
+import { cn } from '@/lib/utils';
 import * as TogglePrimitive from '@radix-ui/react-toggle';
 import { type VariantProps, cva } from 'class-variance-authority';
 import * as React from 'react';
